Cards show, need to reformat, need to do unknown cards
